Mandatory 10s cache for all content
Nginx configuration tweak to force-cache all content for 10s. Tests
on localhost with httperf show an increase in requests served (for a single URL)
from  ~500 /s to several thousand per second.

A value of 10s should be low enough to not really impact usability of anything, due
to forced serving of stale content.

Furthermore, the cache settings are set so that cached content
is served in case the node.js instance is unreachable, takes longer than 3s to respond
or returns with status code 500, 502, 503, 504, or is concurrently serving another
request for the same URL.

It might be wise to deprioritize / nice / cgroup the node.js process compared to nginx,
to have a better chance at keeping the site responsible and serving stale files
in case it get DDOS-attacked.