Revert "Skip IO operations in case there's non binaries update (#239)"
This reverts commit 5010a746de1d118ac6e8b0573fcb475b6f5e4f86.

For some strange reason it seems like the npm build process nuke
the binaries every time is run, that means that with the change we
are reverting we were not copying binaries from the binaries
git repo to the filesystem exposed by the webserver, if no changes were
detected during the "git pull" over the binaries repo.

Hence everyone trying to download would have obtained a 404 error.